fixing some bugs

+++ BizData/Entities/AccionC_Material.cs
@@ -7,7 +7,7 @@ namespace BizData.Entities
     public class AccionC_Material
     {
         public int AccionC_MaterialID { get; set; }
-        public Decimal Precio { get; set; }
+        public double Precio { get; set; }
 
         public virtual AccionConstructiva AccionConstructiva { get; set; }
         public virtual Material Material { get; set; }
